+++ zarrs/tests/zarr_python_compat.rs
@@ -1,12 +1,14 @@
 use std::{error::Error, path::PathBuf, sync::Arc};
 
 use zarrs::{array::Array, array_subset::ArraySubset};
-use zarrs_storage::{storage_adapter::zip::ZipStorageAdapter, store, StoreKey};
+use zarrs_filesystem::FilesystemStore;
+use zarrs_storage::StoreKey;
+use zarrs_zip::ZipStorageAdapter;
 
 #[test]
 fn zarr_python_compat_zip_store() -> Result<(), Box<dyn Error>> {
     let path = PathBuf::from("tests/data/zarr_python_compat");
-    let store = Arc::new(store::FilesystemStore::new(&path)?);
+    let store = Arc::new(FilesystemStore::new(&path)?);
     let store = Arc::new(ZipStorageAdapter::new(store, StoreKey::new("zarr.zip")?)?);
 
     let array = Array::open(store, "/foo")?;