+++ zarrs_storage/src/storage_adapter.rs
@@ -2,8 +2,5 @@
 //!
 //! Storage adapters can be layered on stores.
 
-#[cfg(feature = "zip")]
-pub mod zip;
-
 #[cfg(feature = "async")]
 pub mod async_to_sync;