+++ zarrs_storage/src/store/store_sync.rs
@@ -1,5 +1,0 @@
-pub mod filesystem_store;
-pub mod memory_store;
-
-#[cfg(feature = "http")]
-pub mod http_store;